# api/openapi.yaml
openapi: 3.1.0
info:
  title: SerpentSeek API
  version: "1.0.0"
  summary: Search gateway microservice — visual provider chains, live request tracing, API-key/Passkeys auth, SQLite/PostgreSQL storage.
  description: |
    SerpentSeek is a drop-in replacement for the serpent-shim used by Open WebUI, upgraded to a full
    service with an admin SPA.

    **Machine endpoint.** `POST /search` is Open WebUI-compatible: request `{query, count}`, response is always
    HTTP 200 with an array of `[{link,title,snippet}]` (or `[]` on failure/empty). Diagnostics are carried via
    response headers `X-Serpent-Rid` and `X-Serpent-Api` and the History UI.

    **Auth.** Two mechanisms, either or both can be used:
    - `bearerAuth`: API keys (`seek_ak_…` scope `search`; `seek_admin_…` scope `admin`; `read` for GET-only).
    - `cookieAuth`: session cookie `seek_sid` + CSRF double-submit header `X-CSRF-Token` on all mutations.
    When `AUTH_ENABLED=false` every `/api/*` endpoint behaves as open-admin and `/search` needs no key.

    **Request tracing.** Every response carries `X-Serpent-Rid`. Live step-by-step execution of a request's
    chain is exposed over SSE (see `/api/requests/{id}/events`).

    **Errors.** All `4xx/5xx` responses: `{ "error": { "code", "message", "rid" } }`.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local default (docker compose up --build)
...
security:
  - bearerAuth: []
  - cookieAuth: []
    csrfHeader: []
paths:
...
  /search:
    post:
      tags: [machine]
      summary: Drop-in search endpoint for Open WebUI (replaces serpent-shim)
      description: |
        Compatible with the serpent-shim contract: Bearer auth, `{query, count}`, HTTP 200 **always** —
        either `Row[]` (≤ count) or `[]` on failure/empty. Who answered and the request id are returned in
        headers; full diagnostics live in History (`/api/requests/{id}`).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query: { type: string, minLength: 1, description: Search text }
                count:
                  type: integer
                  minimum: 1
                  maximum: 20
                  default: 5
                  description: Max rows to return
            example: { query: "open webui searxng json", count: 5 }
      responses:
        "200":
          description: Search rows (empty array means fail/empty — see History and response headers)
          headers:
            X-Serpent-Rid: { $ref: "#/components/headers/XSerpentRid" }
            X-Serpent-Api: { $ref: "#/components/headers/XSerpentApi" }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Row" }
              example:
                - link: "https://searxng.org/"
                  title: "SearXNG"
                  snippet: "Privacy-respecting metasearch engine."
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ---------------- auth ----------------
...
  /api/auth/setup:
    post:
      tags: [auth]
      summary: First-run setup (creates first admin + first API key)
      description: Available only while no admin exists; requires the one-time `SETUP_TOKEN` printed to server logs.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [setup_token, admin_name]
              properties:
                setup_token: { type: string }
                admin_name: { type: string }
                create_key: { type: boolean, default: true }
                key_name: { type: string, default: first-admin-key }
      responses:
        "201":
          description: Admin created; API key secret is shown once
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  api_key: { $ref: "#/components/schemas/ApiKeyCreated" }
        "409": { $ref: "#/components/responses/Conflict" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "410":
          description: Setup window closed (token expired or setup already done)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /api/auth/login:
    post:
      tags: [auth]
      summary: Login with an API key (creates a session cookie)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [api_key]
              properties: { api_key: { type: string, example: "seek_ak_ab12cd34.…" } }
      responses:
        "200":
          description: Session established; `seek_sid` (HttpOnly) + `seek_csrf` cookies set
          headers: { X-Serpent-Rid: { $ref: "#/components/headers/XSerpentRid" } }
          content:
            application/json:
              schema:
                type: object
                properties: { user: { $ref: "#/components/schemas/User" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /api/auth/passkey/register/begin:
    post:
      tags: [auth]
      summary: WebAuthn registration — begin (requires an authenticated session)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties: { name: { type: string, description: Friendly name for the passkey } }
      responses:
        "200":
          description: PublicKeyCredentialCreationOptions (webauthn protocol object)
          content:
            application/json:
              schema: { type: object, description: navigator.credentials.create() options }
  /api/auth/passkey/register/finish:
    post:
      tags: [auth]
      summary: WebAuthn registration — finish (stores the credential)
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, description: Attestation response from the authenticator }
      responses:
        "201":
          description: Credential stored
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PasskeyCredential" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /api/auth/passkey/assert/finish:
    post:
      tags: [auth]
      summary: WebAuthn login — finish (creates session cookie)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, description: Assertion response from the authenticator }
      responses:
        "200":
          description: Session established
          content:
            application/json:
              schema:
                type: object
                properties: { user: { $ref: "#/components/schemas/User" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }

  # ---------------- users / keys / passkeys ----------------
  /api/users:
    get:
      tags: [users]
      summary: List users (admin)
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
    post:
      tags: [users]
      summary: Create user (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/users/{id}:
    patch:
      tags: [users]
      summary: Update user (rename, role, disable)
      parameters: [ { $ref: "#/components/parameters/UserId" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserPatch" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [users]
      summary: Delete user (admin); own admin account cannot delete itself while it is the last admin
      parameters: [ { $ref: "#/components/parameters/UserId" } ]
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/keys:
    get:
      tags: [users]
      summary: List API keys (prefix only — secrets are never returned)
      responses:
        "200":
          description: API keys
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ApiKey" }
    post:
      tags: [users]
      summary: Create API key (secret returned ONCE)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApiKeyCreateInput" }
      responses:
        "201":
          description: Created — copy the secret now, it is never shown again
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiKeyCreated" }
  /api/keys/{id}:
    patch:
      tags: [users]
      summary: Update key (rename, expiry, revoke)
      parameters: [ { $ref: "#/components/parameters/KeyId" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                expires_at: { type: string, format: date-time, nullable: true }
                revoke: { type: boolean }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiKey" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [users]
      summary: Delete key permanently
      parameters: [ { $ref: "#/components/parameters/KeyId" } ]
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/passkeys/{id}:
    delete:
      tags: [users]
      summary: Delete a passkey credential
      parameters: [ { $ref: "#/components/parameters/UserId" } ]
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------------- history ----------------
  /api/requests:
    get:
      tags: [history]
      summary: Search request history (cursor pagination, filters)
      parameters:
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/RequestStatus" }
        - name: provider
          in: query
          schema: { type: string, description: Filter by step provider code }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - name: q
          in: query
          schema: { type: string, description: Substring match on query text }
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
      responses:
        "200":
          description: Page of requests
          headers: { X-Serpent-Rid: { $ref: "#/components/headers/XSerpentRid" } }
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/RequestSummary" }
                  next_cursor: { type: string, nullable: true }
  /api/requests/{id}:
    get:
      tags: [history]
      summary: Request details + step trace + chain snapshot (graph as it was at run time)
      parameters: [ { $ref: "#/components/parameters/RequestId" } ]
      responses:
        "200":
          description: Full request record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RequestDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/requests/{id}/events:
    get:
      tags: [history]
      summary: Live execution trace of a request (SSE)
      description: |
        Server-Sent Events stream. Wire format: `event: <name>\ndata: <json>\n\nid: <seq>`.
        Server sends a keep-alive ping every 15s and sets `retry: 2000`. Reconnection is supported via
        the `Last-Event-ID` header. If the request already finished, the server immediately replays all
        stored steps and closes the stream with `request_done`.

        Events:
        - `step_started` → SSEStepStarted — a chain node began executing (🟡 in UI)
        - `step_finished` → SSEStepFinished — node completed (🟢 ok / 🔴 fail / ⚪ skip: see `kind` and `status`)
        - `provider_dead` → SSEProviderDead — provider is excluded for the rest of THIS request
        - `merge_progress` → SSEMergeProgress — full-chain live counter (collected/unique links so far)
        - `merge_done` → SSEMergeDone — full-chain terminal merge counters (before `request_done`)
        - `request_done` → SSERequestDone — terminal event; stream closes right after

        The UI maps steps onto the chain snapshot: running node = yellow pulse, ok = green, fail = red,
        skipped = gray; ProgressCube fills with each finished step and turns red on abort.
      parameters: [ { $ref: "#/components/parameters/RequestId" } ]
      responses:
        "200":
          description: SSE stream
          headers:
            Cache-Control: { schema: { type: string }, description: "no-cache" }
            Last-Event-ID:
              schema: { type: string }
              description: Echo of the client's last received event id (request header)
          content:
            text/event-stream:
              schema: { type: string, description: SSE wire format; see operation description for event payloads }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/chains:
    get:
      tags: [chains]
      summary: List chains (summaries)
      responses:
        "200":
          description: Chains
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ChainSummary" }
    post:
      tags: [chains]
      summary: Create chain (nodes + edges + layout); server validates the graph
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChainInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Chain" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /api/chains/{id}:
    get:
      tags: [chains]
      summary: Chain details (graph + layout)
      parameters: [ { $ref: "#/components/parameters/ChainId" } ]
      responses:
        "200":
          description: Chain
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Chain" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [chains]
      summary: Replace graph (nodes/edges/positions) + metadata; validated before save
      parameters: [ { $ref: "#/components/parameters/ChainId" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChainInput" }
      responses:
        "200":
          description: Updated (version is incremented)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Chain" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [chains]
      summary: Delete chain (the ACTIVE chain cannot be deleted)
      parameters: [ { $ref: "#/components/parameters/ChainId" } ]
      responses:
        "204": { description: Deleted }
        "409": { $ref: "#/components/responses/Conflict" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/chains/{id}/activate:
    post:
      tags: [chains]
      summary: Make chain active (exactly one active chain at a time)
      parameters: [ { $ref: "#/components/parameters/ChainId" } ]
      responses:
        "200":
          description: Activated
          content:
            application/json:
              schema:
                type: object
                properties: { ok: { type: boolean } }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/chains/{id}/duplicate:
    post:
      tags: [chains]
      summary: Duplicate chain ("<name> copy")
      parameters: [ { $ref: "#/components/parameters/ChainId" } ]
      responses:
        "201":
          description: Duplicated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Chain" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/providers:
    get:
      tags: [providers]
      summary: 'List providers (credentials masked: only `"***"` + set flag)'
      responses:
        "200":
          description: Providers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Provider" }
    post:
      tags: [providers]
      summary: Create provider (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProviderInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Provider" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/providers/{id}:
    get:
      tags: [providers]
      summary: Provider details
      parameters: [ { $ref: "#/components/parameters/ProviderId" } ]
      responses:
        "200":
          description: Provider
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Provider" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [providers]
      summary: Full replace of a provider
      parameters: [ { $ref: "#/components/parameters/ProviderId" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProviderInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Provider" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [providers]
      summary: Partial update — enabled toggle and/or base_url
      description: |
        Used by the card on/off switch (`enabled`) and the URL field (`base_url`).
        `base_url` is **not a secret** (e.g. the external SearXNG URL) and is freely
        readable/editable. Full edits (name/params/credentials) go through `PUT`.
      parameters: [ { $ref: "#/components/parameters/ProviderId" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProviderPatch" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Provider" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [providers]
      summary: Delete a provider instance (refused while a chain references it)
      parameters: [ { $ref: "#/components/parameters/ProviderId" } ]
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties: { ok: { type: boolean } }
        "409":
          description: Provider is used by one or more chains (`error.chains[]`); remove it from the chains first
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/providers/{id}/test:
    post:
      tags: [providers]
      summary: Live test of a provider (probe query "test"; shows http/kind/ms and up to 3 links)
      parameters: [ { $ref: "#/components/parameters/ProviderId" } ]
      responses:
        "200":
          description: Probe result (a failed probe is HTTP 200 with ok=false — classification is in `kind`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProviderTestResult" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------------- proxies ----------------
  /api/proxies:
    get:
      tags: [providers]
      summary: List named outbound proxies (password values are never returned)
      responses:
        "200":
          description: Proxies
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Proxy" }
    post:
      tags: [providers]
      summary: Create a named proxy (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProxyInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Proxy" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /api/settings:
    get:
      tags: [settings]
      summary: All settings (with source env|db|default and readonly flags)
      responses:
        "200":
          description: Settings catalog
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SettingsResponse" }
    put:
      tags: [settings]
      summary: Update settings (db-sourced only; env-managed keys are readonly)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SettingsUpdateRequest" }
      responses:
        "200":
          description: Applied (some keys need a container restart — e.g. STORAGE_DRIVER)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SettingsUpdateResponse" }
        "422": { $ref: "#/components/responses/ValidationFailed" }

  # ---------------- logs ----------------
...
  /api/search/ui:
    post:
      tags: [playground]
      summary: Start a manual search from the UI (async — trace it live over SSE)
      description: Creates the request and returns immediately; the UI subscribes to `/api/requests/{id}/events`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "202":
          description: Accepted — follow the trace at `stream_url`
          headers: { X-Serpent-Rid: { $ref: "#/components/headers/XSerpentRid" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlaygroundResponse" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        API key in `Authorization: Bearer …`. Scopes: `search` → `/search` only; `read` → GET `/api/*`;
        `admin` → everything. Key format `seek_ak_<prefix>.<secret>` / `seek_admin_…`.
    cookieAuth:
      type: apiKey
      in: cookie
      name: seek_sid
      description: Session cookie (HttpOnly, Secure, SameSite=Strict, rolling expiry 30d).
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Double-submit CSRF token (readable cookie `seek_csrf` → echo in this header) — required on all mutations made with cookieAuth.
...